refactor(core): delete IntegrationRenderer indirection wrappers
Make page-browser-graph merge pure, finalize document shells with caller-owned inputs, and have integrations call ForeignSubtreeExecutionService directly. Redistribute coordinator tests by domain.
