+++ packages/core/jsr.json
@@ -9,7 +9,9 @@
 		"./route-renderer/orchestration/string-markup-renderer": "./src/route-renderer/orchestration/string-markup-renderer.ts",
 		"./route-renderer/orchestration/foreign-child/component-render-context": "./src/route-renderer/orchestration/foreign-child/component-render-context.ts",
 		"./route-renderer/orchestration/foreign-child/foreign-subtree-execution.service": "./src/route-renderer/orchestration/foreign-child/foreign-subtree-execution.service.ts",
+		"./route-renderer/orchestration/foreign-child/owning-renderer-resolution": "./src/route-renderer/orchestration/foreign-child/owning-renderer-resolution.ts",
 		"./route-renderer/orchestration/document-shell/document-shell-render.service": "./src/route-renderer/orchestration/document-shell/document-shell-render.service.ts",
+		"./route-renderer/orchestration/page-browser-graph/page-browser-graph-merge.utils": "./src/route-renderer/orchestration/page-browser-graph/page-browser-graph-merge.utils.ts",
 		"./errors/locals-access-error": "./src/errors/locals-access-error.ts",
 		"./constants": "./src/constants.ts",
 		"./config-builder": "./src/config/config-builder.ts",
