+++ packages/core/src/route-renderer/orchestration/integration-renderer.test.ts
@@ -1,244 +1,28 @@
 import { describe, it, expect, vi } from 'vitest';
-import { eco } from '../../eco/eco.ts';
 import { createPagePackage } from '../../services/assets/asset-processing-service/index.ts';
-import {
-	IntegrationRenderer,
-	type HtmlDocumentContribution,
-	type RenderToResponseContext,
-} from './integration-renderer.ts';
-import { resolveInnermostPageLayout } from './document-shell/layout-shell-props.service.ts';
+import { IntegrationRenderer } from './integration-renderer.ts';
 import type { EcoPagesAppConfig } from '../../types/internal-types.ts';
 import type { AssetProcessingService, ProcessedAsset } from '../../services/assets/asset-processing-service/index.ts';
-import { getComponentRenderContext, type ForeignChildRuntime } from './foreign-child/component-render-context.ts';
 import type {
-	BaseIntegrationContext,
 	ComponentRenderInput,
-	ComponentRenderResult,
-	ForeignSubtreeRenderPayload,
-	RouteRendererBody,
 	EcoPagesElement,
 	IntegrationRendererRenderOptions,
 	EcoPageFile,
-	PageBrowserGraphContribution,
+	RouteRendererBody,
 	RouteRendererOptions,
 	EcoComponent,
 	HtmlTemplateProps,
+	PageBrowserGraphContribution,
 } from '../../types/public-types.ts';
 import type { EcoPageComponent } from '../../eco/eco.types.ts';
-import { runWithComponentRenderContext } from './foreign-child/component-render-context.ts';
-import { toForeignSubtreeRenderPayload } from './foreign-child/foreign-subtree-execution.service.ts';
-
-function createUnresolvedMarkerArtifact(nodeId: string, componentRef: string, propsRef: string): string {
-	return `<eco-marker data-eco-node-id="${nodeId}" data-eco-component-ref="${componentRef}" data-eco-props-ref="${propsRef}"></eco-marker>`;
-}
-
-function createMockIntegrationPlugin(
-	plugin: { name: string } & Record<string, unknown>,
-): EcoPagesAppConfig['integrations'][number] {
-	return {
-		setConfig: vi.fn(),
-		setRuntimeOrigin: vi.fn(),
-		setup: vi.fn(async () => {}),
-		...plugin,
-	} as unknown as EcoPagesAppConfig['integrations'][number];
-}
-
-/**
- * Concrete implementation with ed file loading for testing purposes.
- */
-class TestIntegrationRenderer extends IntegrationRenderer<EcoPagesElement> {
-	name = 'test-renderer';
-	ForeignChildRuntimeCreationCount = 0;
-	ForeignChildRenderCount = 0;
-	UseFailFastForeignChildRuntime = false;
-
-	/** Mock data container for page module */
-	PageModule: EcoPageFile | null = null;
-	/** Mock HTML template container */
-	HtmlTemplate: EcoComponent<HtmlTemplateProps> | null = null;
-	RenderedBody: RouteRendererBody = '<html><body>Test Page</body></html>';
-	RenderBodyFactory:
-		| ((context: ReturnType<typeof getComponentRenderContext>) => RouteRendererBody | Promise<RouteRendererBody>)
-		| null = null;
-	MockComponentRenderResult: ComponentRenderResult | null = null;
-	ImportedFiles: string[] = [];
-	HtmlContributions: HtmlDocumentContribution[] = [];
-
-	async render(_options: IntegrationRendererRenderOptions<EcoPagesElement>): Promise<RouteRendererBody> {
-		if (this.RenderBodyFactory) {
-			return await this.RenderBodyFactory(getComponentRenderContext());
-		}
-
-		return this.RenderedBody;
-	}
-
-	override async renderComponent(_input: ComponentRenderInput): Promise<ComponentRenderResult> {
-		if (this.MockComponentRenderResult) {
-			return this.MockComponentRenderResult;
-		}
-
-		return super.renderComponent(_input);
-	}
-
-	override async renderComponentWithForeignChildren(input: ComponentRenderInput): Promise<ComponentRenderResult> {
-		this.ForeignChildRenderCount += 1;
-		return await super.renderComponentWithForeignChildren(input);
-	}
-
-	async renderToResponse<P>(view: EcoComponent<P>, props: P, ctx: RenderToResponseContext): Promise<Response> {
-		const viewFn = view as (props: P) => EcoPagesElement;
-		const content = viewFn(props);
-
-		let body: string;
-		if (ctx.partial) {
-			body = content as string;
-		} else {
-			const Layout = resolveInnermostPageLayout(view.config?.layouts) as
-				((props: { children: string }) => string) | undefined;
-			const children = Layout ? Layout({ children: content as string }) : content;
-			body = `<!DOCTYPE html><html><body>${children}</body></html>`;
-		}
-
-		const headers = new Headers({ 'Content-Type': 'text/html; charset=utf-8' });
-		if (ctx.headers) {
-			for (const [key, value] of Object.entries(ctx.headers)) {
-				headers.set(key, value);
-			}
-		}
-
-		return new Response(body, {
-			status: ctx.status ?? 200,
-			headers,
-		});
-	}
-
-	/**
-	 * Override protected methods to return data.
-	 */
-	protected override async importPageFile(_file: string): Promise<EcoPageFile> {
-		this.ImportedFiles.push(_file);
-		if (!this.PageModule) throw new Error('Mock page module not set');
-		return this.PageModule;
-	}
-
-	protected override async getHtmlTemplate(): Promise<EcoComponent<HtmlTemplateProps>> {
-		if (!this.HtmlTemplate) throw new Error('Mock HTML template not set');
-		return this.HtmlTemplate;
-	}
-
-	/**
-	 * Override to avoid asset processing service dependency in tests.
-	 */
-	protected override async resolveDependencies(
-		_components: (EcoComponent | Partial<EcoComponent>)[],
-	): Promise<ProcessedAsset[]> {
-		return [];
-	}
-
-	/**
-	 * Expose protected method for testing.
-	 */
-	public async testPrepareRenderOptions(options: RouteRendererOptions) {
-		return this.prepareRenderOptions(options);
-	}
-
-	public testShouldDelegateForeignChild(input: { currentIntegration: string; targetIntegration?: string }) {
-		return this.foreignSubtreeExecutionService.shouldDelegateForeignChild(input);
-	}
-
-	public testHasForeignChildDescendants(component: EcoComponent) {
-		return this.hasForeignChildDescendants(component);
-	}
-
-	public async testGetHtmlTemplate() {
-		return this.getHtmlTemplate();
-	}
-
-	public async testBaseGetHtmlTemplate() {
-		return super.getHtmlTemplate();
-	}
-
-	public testGetRendererBootstrapDependencies(partial = false) {
-		return this.getRendererBootstrapDependencies(partial);
-	}
-
-	public async testFinalizeResolvedHtml(options: { html: string; partial?: boolean }) {
-		return this.finalizeResolvedHtml({
-			html: options.html,
-			partial: options.partial,
-		});
-	}
-
-	public async testRenderPartialViewResponse<P>(input: {
-		view: EcoComponent<P>;
-		props: P;
-		ctx?: RenderToResponseContext;
-		renderInline?: () => Promise<BodyInit>;
-		transformHtml?: (html: string) => string;
-	}) {
-		return this.renderPartialViewResponse({
-			view: input.view,
-			props: input.props,
-			ctx: input.ctx ?? { partial: true },
-			renderInline: input.renderInline,
-			transformHtml: input.transformHtml,
-		});
-	}
-
-	public async testRenderViewWithDocumentShell<P>(input: {
-		view: EcoComponent<P>;
-		props: P;
-		ctx?: RenderToResponseContext;
-		layout?: EcoComponent;
-	}) {
-		return this.renderViewWithDocumentShell({
-			view: input.view,
-			props: input.props,
-			ctx: input.ctx ?? {},
-			layout: input.layout,
-		});
-	}
-
-	public async testRenderForeignSubtree(input: ComponentRenderInput) {
-		return toForeignSubtreeRenderPayload(await this.renderComponentWithForeignChildren(input));
-	}
-
-	protected override createForeignChildRuntime(options: {
-		renderInput: ComponentRenderInput;
-		rendererCache: Map<string, IntegrationRenderer<any>>;
-	}): ForeignChildRuntime {
-		this.ForeignChildRuntimeCreationCount += 1;
-
-		if (this.UseFailFastForeignChildRuntime) {
-			return this.createFailFastForeignChildRuntime();
-		}
-
-		return super.createForeignChildRuntime(options);
-	}
-
-	protected override getHtmlDocumentContributions(): HtmlDocumentContribution[] | undefined {
-		return this.HtmlContributions;
-	}
-}
+import {
+	TestIntegrationRenderer,
+	createMockIntegrationPlugin,
+	testAppConfig,
+	testAssetService,
+} from './integration-renderer.test-fixtures.ts';
 
 describe('IntegrationRenderer', () => {
-	const AppConfig = {
-		absolutePaths: {
-			pagesDir: '/app/pages',
-			htmlTemplatePath: '/app/index.ghtml.ts',
-		},
-		integrations: [],
-		defaultMetadata: {
-			title: 'Default Title',
-			description: 'Default Description',
-		},
-		srcDir: '/app',
-	} as unknown as EcoPagesAppConfig;
-
-	const AssetService = {
-		processDependencies: vi.fn(() => Promise.resolve([])),
-	} as unknown as AssetProcessingService;
-
 	it('serializes concurrent execute calls on one renderer instance', async () => {
 		let active = 0;
 		let maxActive = 0;
@@ -255,8 +39,8 @@ describe('IntegrationRenderer', () => {
 				return '<html><body>ok</body></html>';
 			}
 		})({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
+			appConfig: testAppConfig,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 		renderer.PageModule = { default: Page };
@@ -310,7 +94,7 @@ describe('IntegrationRenderer', () => {
 				return await super.prepareRenderOptions(options);
 			}
 		})({
-			appConfig: AppConfig,
+			appConfig: testAppConfig,
 			assetProcessingService: { processDependencies } as unknown as AssetProcessingService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
@@ -342,8 +126,8 @@ describe('IntegrationRenderer', () => {
 
 	it('transformRouteResponse prefers renderer-owned page package updates over stale prepare-time package', async () => {
 		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
+			appConfig: testAppConfig,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 		const stalePagePackage = createPagePackage([]);
@@ -355,10 +139,10 @@ describe('IntegrationRenderer', () => {
 			attributes: { type: 'module' },
 		} as ProcessedAsset;
 
-		(renderer as any).htmlTransformer.setPagePackage(stalePagePackage);
-		(renderer as any).appendProcessedDependencies([renderTimeAsset]);
+		renderer.setTestPagePackage(stalePagePackage);
+		renderer.appendTestProcessedDependencies([renderTimeAsset]);
 
-		const body = await (renderer as any).transformRouteResponse(
+		const body = await renderer.testTransformRouteResponse(
 			new Response('<html><head></head><body></body></html>', {
 				headers: { 'Content-Type': 'text/html' },
 			}),
@@ -370,58 +154,14 @@ describe('IntegrationRenderer', () => {
 		expect(html).toContain('/assets/scripts/render-time-hydration.js');
 	});
 
-	it('preserves the page browser graph when appending renderer-owned assets', () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-		const pageEntryAsset = {
-			kind: 'script',
-			inline: false,
-			srcUrl: '/page.js',
-			position: 'head',
-			packageRole: 'page-script',
-		} as ProcessedAsset;
-		const rendererAsset = {
-			kind: 'stylesheet',
-			inline: false,
-			srcUrl: '/component.css',
-			position: 'head',
-		} as ProcessedAsset;
-
-		(renderer as any).htmlTransformer.setPagePackage(
-			createPagePackage([], {
-				pageBrowserGraph: {
-					entryAssets: [pageEntryAsset],
-					chunkAssets: [],
-				},
-			}),
-		);
-
-		(renderer as any).appendProcessedDependencies([rendererAsset]);
-
-		expect((renderer as any).htmlTransformer.getPagePackage()).toEqual(
-			expect.objectContaining({
-				pageBrowserGraph: {
-					entryAssets: [pageEntryAsset],
-					chunkAssets: [],
-				},
-				assets: expect.arrayContaining([rendererAsset, pageEntryAsset]),
-			}),
-		);
-	});
-
 	it('should extract cache strategy from page component (static property)', async () => {
 		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
+			appConfig: testAppConfig,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 
-		/** Mock page with cache strategy */
 		const PageIdx = (() => 'Page Content') as EcoPageComponent<any>;
-		/** Simulate eco.page() attaching cache config */
 		PageIdx.cache = { revalidate: 60 };
 
 		renderer.PageModule = {
@@ -429,25 +169,22 @@ describe('IntegrationRenderer', () => {
 		};
 		renderer.HtmlTemplate = (() => 'HTML Template') as EcoComponent<HtmlTemplateProps>;
 
-		const options: RouteRendererOptions = {
+		const result = await renderer.testPrepareRenderOptions({
 			file: '/app/pages/cached-page.ts',
 			params: {},
 			query: {},
-		};
-
-		const result = await renderer.testPrepareRenderOptions(options);
+		});
 
 		expect(result.cacheStrategy).toEqual({ revalidate: 60 });
 	});
 
 	it('should return undefined cache strategy if not present', async () => {
 		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
+			appConfig: testAppConfig,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 
-		/** Mock page without cache strategy */
 		const PageIdx = (() => 'Page Content') as EcoPageComponent<any>;
 
 		renderer.PageModule = {
@@ -455,26 +192,23 @@ describe('IntegrationRenderer', () => {
 		};
 		renderer.HtmlTemplate = (() => 'HTML Template') as EcoComponent<HtmlTemplateProps>;
 
-		const options: RouteRendererOptions = {
+		const result = await renderer.testPrepareRenderOptions({
 			file: '/app/pages/simple-page.ts',
 			params: {},
 			query: {},
-		};
-
-		const result = await renderer.testPrepareRenderOptions(options);
+		});
 
 		expect(result.cacheStrategy).toBeUndefined();
 	});
 
 	it('should resolve static props and metadata correctly', async () => {
 		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
+			appConfig: testAppConfig,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 
 		const PageIdx = (() => 'Page Content') as EcoPageComponent<any>;
-		/** Attached static methods */
 		PageIdx.staticProps = async () => ({ props: { title: 'Dynamic Title' } });
 		PageIdx.metadata = async ({ props }: { props: Record<string, unknown> }) => ({
 			title: props.title as string,
@@ -486,13 +220,11 @@ describe('IntegrationRenderer', () => {
 		};
 		renderer.HtmlTemplate = (() => 'HTML Template') as EcoComponent<HtmlTemplateProps>;
 
-		const options: RouteRendererOptions = {
+		const result = await renderer.testPrepareRenderOptions({
 			file: '/app/pages/props-page.ts',
 			params: {},
 			query: {},
-		};
-
-		const result = await renderer.testPrepareRenderOptions(options);
+		});
 
 		expect(result.props).toEqual({ title: 'Dynamic Title' });
 		expect(result.metadata?.title).toBe('Dynamic Title');
@@ -501,14 +233,14 @@ describe('IntegrationRenderer', () => {
 	it('should prefer renderer module html template path when provided', async () => {
 		const renderer = new TestIntegrationRenderer({
 			appConfig: {
-				...AppConfig,
+				...testAppConfig,
 				runtime: {
 					rendererModuleContext: {
 						htmlTemplateModulePath: '/virtual/includes/html.kita.tsx',
 					},
 				},
 			} as EcoPagesAppConfig,
-			assetProcessingService: AssetService,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 
@@ -521,110 +253,10 @@ describe('IntegrationRenderer', () => {
 		expect(renderer.ImportedFiles).toContain('/virtual/includes/html.kita.tsx');
 	});
 
-	it('should expose island bootstrap dependencies from renderer modules', () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: {
-				...AppConfig,
-				runtime: {
-					rendererModuleContext: {
-						islandClientModuleId: 'virtual:ecopages/island-client.ts',
-					},
-				},
-			} as EcoPagesAppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		expect(renderer.testGetRendererBootstrapDependencies()).toEqual([
-			{
-				attributes: {
-					crossorigin: 'anonymous',
-					'data-ecopages-runtime': 'islands',
-					type: 'module',
-				},
-				content: 'import "virtual:ecopages/island-client.ts";',
-				inline: true,
-				kind: 'script',
-				packageRole: 'keep-separate',
-				position: 'body',
-			},
-		]);
-		expect(renderer.testGetRendererBootstrapDependencies(true)).toEqual([]);
-	});
-
-	it('should inject the island client bootstrap into finalized full-document HTML', async () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: {
-				...AppConfig,
-				runtime: {
-					rendererModuleContext: {
-						islandClientModuleId: 'virtual:ecopages/island-client.ts',
-					},
-				},
-			} as EcoPagesAppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		const html = await renderer.testFinalizeResolvedHtml({
-			html: '<!DOCTYPE html><html><body><main>hello</main></body></html>',
-		});
-
-		expect(html).toContain('data-ecopages-runtime="islands"');
-		expect(html).toContain('import "virtual:ecopages/island-client.ts";');
-	});
-
-	it('should inject declarative html contributions during final html transformation', async () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-		renderer.HtmlContributions = [
-			{ placement: 'head-prepend', html: '<meta name="integration-slot" content="head">' },
-			{ placement: 'body-append', html: '<script id="integration-tail">window.tail = true;</script>' },
-		];
-
-		const html = await renderer.testFinalizeResolvedHtml({
-			html: '<!DOCTYPE html><html><head><title>Page</title></head><body><main>hello</main></body></html>',
-		});
-
-		expect(html).toContain('<head><meta name="integration-slot" content="head"><title>Page</title>');
-		expect(html).toContain('<script id="integration-tail">window.tail = true;</script></body>');
-	});
-
-	it('should inject declarative html contributions during shared route execution', async () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-		renderer.PageModule = {
-			default: (() => 'Hello Route') as EcoPageComponent<any>,
-		};
-		renderer.HtmlTemplate = (() =>
-			'<html><head><title>Route</title></head><body><main>Hello Route</main></body></html>') as EcoComponent<HtmlTemplateProps>;
-		renderer.RenderedBody = '<html><head><title>Route</title></head><body><main>Hello Route</main></body></html>';
-		renderer.HtmlContributions = [
-			{ placement: 'head-prepend', html: '<meta name="route-slot" content="head">' },
-			{ placement: 'body-append', html: '<script id="route-tail">window.routeTail = true;</script>' },
-		];
-
-		const result = await renderer.execute({
-			file: '/app/pages/route.tsx',
-			params: {},
-			query: {},
-		});
-		const html = await new Response(result.body as BodyInit).text();
-
-		expect(html).toContain('<head><meta name="route-slot" content="head"><title>Route</title>');
-		expect(html).toContain('<script id="route-tail">window.routeTail = true;</script></body>');
-	});
-
 	it('should keep layout locals safe and page locals guarded on static pages', async () => {
 		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
+			appConfig: testAppConfig,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 
@@ -647,8 +279,8 @@ describe('IntegrationRenderer', () => {
 
 	it('should provide both locals and pageLocals on dynamic pages', async () => {
 		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
+			appConfig: testAppConfig,
+			assetProcessingService: testAssetService,
 			runtimeOrigin: 'http://localhost:3000',
 		});
 
@@ -672,367 +304,11 @@ describe('IntegrationRenderer', () => {
 		expect(result.pageLocals).toBe(incomingLocals);
 	});
 
-	it('should prepare render options when declared foreign dependencies are valid', async () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: {
-				...AppConfig,
-				integrations: [
-					{
-						name: 'foreign-renderer',
-					} as unknown as EcoPagesAppConfig['integrations'][number],
-				],
-			} as EcoPagesAppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		const ForeignComponent = (() => '<aside>Foreign</aside>') as EcoComponent<Record<string, unknown>>;
-		ForeignComponent.config = {
-			integration: 'foreign-renderer',
-			__eco: {
-				id: 'foreign-component',
-				file: '/app/components/foreign-component.tsx',
-				integration: 'foreign-renderer',
-			},
-		};
-
-		const Layout = (() => '<main>Layout</main>') as EcoComponent<Record<string, unknown>>;
-		Layout.config = {
-			__eco: {
-				id: 'layout-component',
-				file: '/app/layouts/default.tsx',
-				integration: 'test-renderer',
-			},
-			dependencies: {
-				components: [ForeignComponent],
-			},
-		};
-
-		const PageIdx = (() => 'Page Content') as EcoPageComponent<any>;
-		PageIdx.config = {
-			layouts: [Layout],
-			__eco: {
-				id: 'page-component',
-				file: '/app/pages/index.tsx',
-				integration: 'test-renderer',
-			},
-		};
-
-		const HtmlTemplate = (() => 'HTML Template') as EcoComponent<HtmlTemplateProps>;
-		HtmlTemplate.config = {
-			__eco: {
-				id: 'html-template',
-				file: '/app/index.ghtml.ts',
-				integration: 'test-renderer',
-			},
-		};
-
-		renderer.PageModule = {
-			default: PageIdx,
-		};
-		renderer.HtmlTemplate = HtmlTemplate;
-
-		await expect(
-			renderer.testPrepareRenderOptions({
-				file: '/app/pages/index.tsx',
-				params: {},
-				query: {},
-			}),
-		).resolves.toEqual(
-			expect.objectContaining({
-				Page: PageIdx,
-				Layout,
-				HtmlTemplate,
-			}),
-		);
-	});
-
-	it('should fail fast for unknown foreign integration owners', async () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		const UnknownForeignComponent = (() => '<aside>Foreign</aside>') as EcoComponent<Record<string, unknown>>;
-		UnknownForeignComponent.config = {
-			integration: 'missing-renderer',
-			__eco: {
-				id: 'missing-foreign-component',
-				file: '/app/components/missing-foreign-component.tsx',
-				integration: 'missing-renderer',
-			},
-		};
-
-		const PageIdx = (() => 'Page Content') as EcoPageComponent<any>;
-		PageIdx.config = {
-			__eco: {
-				id: 'page-component',
-				file: '/app/pages/index.tsx',
-				integration: 'test-renderer',
-			},
-			dependencies: {
-				components: [UnknownForeignComponent],
-			},
-		};
-
-		const HtmlTemplate = (() => 'HTML Template') as EcoComponent<HtmlTemplateProps>;
-		HtmlTemplate.config = {
-			__eco: {
-				id: 'html-template',
-				file: '/app/index.ghtml.ts',
-				integration: 'test-renderer',
-			},
-		};
-
-		renderer.PageModule = {
-			default: PageIdx,
-		};
-		renderer.HtmlTemplate = HtmlTemplate;
-
-		await expect(
-			renderer.testPrepareRenderOptions({
-				file: '/app/pages/index.tsx',
-				params: {},
-				query: {},
-			}),
-		).rejects.toThrow(
-			'[ecopages] Foreign child "missing-foreign-component" references unknown integration owner "missing-renderer".',
-		);
-	});
-
-	it('should expose a compatibility foreign-subtree payload contract', async () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		renderer.MockComponentRenderResult = {
-			html: '<main data-root="true">Hello</main>',
-			canAttachAttributes: true,
-			rootTag: 'main',
-			integrationName: 'test-renderer',
-			rootAttributes: { 'data-eco-component-id': 'root-1' },
-			assets: [
-				{
-					kind: 'script',
-					inline: true,
-					content: 'console.log("foreign-subtree")',
-					position: 'body',
-				},
-			],
-		};
-
-		const payload = await renderer.testRenderForeignSubtree({
-			component: (() => '<main>Hello</main>') as EcoComponent<Record<string, unknown>>,
-			props: {},
-		});
-
-		expect(payload).toEqual<ForeignSubtreeRenderPayload>({
-			html: '<main data-root="true">Hello</main>',
-			assets: [
-				{
-					kind: 'script',
-					inline: true,
-					content: 'console.log("foreign-subtree")',
-					position: 'body',
-				},
-			],
-			rootTag: 'main',
-			rootAttributes: { 'data-eco-component-id': 'root-1' },
-			attachmentPolicy: { kind: 'first-element' },
-			integrationName: 'test-renderer',
-		});
-	});
-
-	it('should resolve foreign-owned boundaries in the owning renderer', () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		const result = renderer.testShouldDelegateForeignChild({
-			currentIntegration: 'ghtml',
-			targetIntegration: 'react',
-		});
-
-		expect(result).toBe(true);
-	});
-
-	it('should keep same-integration boundaries in the current render pass', () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: AppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		const result = renderer.testShouldDelegateForeignChild({
-			currentIntegration: 'react',
-			targetIntegration: 'react',
-		});
-
-		expect(result).toBe(false);
-	});
-
-	it('should delegate foreign component boundaries through the shared execution seam', async () => {
-		const foreignRenderer = {
-			name: 'foreign-renderer',
-			renderComponentWithForeignChildren: vi.fn(async () => ({
-				html: '<aside>Owned by foreign renderer</aside>',
-				canAttachAttributes: true,
-				rootTag: 'aside',
-				integrationName: 'foreign-renderer',
-			})),
-		} as unknown as IntegrationRenderer;
-
-		const initializeRenderer = vi.fn(() => foreignRenderer);
-		const setup = vi.fn(async () => {});
-		const renderer = new TestIntegrationRenderer({
-			appConfig: {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'foreign-renderer',
-						initializeRenderer,
-						setup,
-					}),
-				],
-			} as EcoPagesAppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		const ForeignComponent = (() => '<aside>Foreign</aside>') as EcoComponent<Record<string, unknown>>;
-		ForeignComponent.config = {
-			integration: 'foreign-renderer',
-			__eco: {
-				id: 'foreign-component',
-				file: '/app/components/foreign-component.tsx',
-				integration: 'foreign-renderer',
-			},
-		};
-
-		const rendererCache = new Map<string, IntegrationRenderer<any>>();
-		const result = await renderer.renderComponentWithForeignChildren({
-			component: ForeignComponent,
-			props: { label: 'foreign' },
-			integrationContext: { rendererCache },
-		});
-
-		expect(result).toEqual(
-			expect.objectContaining({
-				html: '<aside>Owned by foreign renderer</aside>',
-				integrationName: 'foreign-renderer',
-			}),
-		);
-		expect(setup).toHaveBeenCalledTimes(1);
-		expect(initializeRenderer).toHaveBeenCalledTimes(1);
-		expect(foreignRenderer.renderComponentWithForeignChildren).toHaveBeenCalledTimes(1);
-	});
-
-	it('should preserve shared integration context fields when delegating to the owning renderer', async () => {
-		const foreignRenderer = {
-			renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) => ({
-				html: `<aside>${String(
-					(input.integrationContext as BaseIntegrationContext | undefined)?.componentInstanceId ?? 'missing',
-				)}</aside>`,
-				canAttachAttributes: true,
-				rootTag: 'aside',
-				integrationName: 'foreign-renderer',
-			})),
-		} as unknown as IntegrationRenderer;
-		const initializeRenderer = vi.fn(() => foreignRenderer);
-
-		const renderer = new TestIntegrationRenderer({
-			appConfig: {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'foreign-renderer',
-						initializeRenderer,
-					}),
-				],
-			} as EcoPagesAppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-
-		const ForeignComponent = (() => '<aside>Foreign</aside>') as EcoComponent<Record<string, unknown>>;
-		ForeignComponent.config = {
-			integration: 'foreign-renderer',
-			__eco: {
-				id: 'foreign-component',
-				file: '/app/components/foreign-component.tsx',
-				integration: 'foreign-renderer',
-			},
-		};
-
-		await renderer.renderComponentWithForeignChildren({
-			component: ForeignComponent,
-			props: { label: 'foreign' },
-			integrationContext: {
-				componentInstanceId: 'host-1',
-			} satisfies BaseIntegrationContext,
-		});
-
-		expect(foreignRenderer.renderComponentWithForeignChildren).toHaveBeenCalledWith(
-			expect.objectContaining({
-				integrationContext: expect.objectContaining({
-					componentInstanceId: 'host-1',
-					rendererCache: expect.any(Map),
-				}),
-			}),
-		);
-	});
-
-	it('should fall back to local rendering when the resolved owner renderer is the current renderer', async () => {
-		const renderer = new TestIntegrationRenderer({
-			appConfig: {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'foreign-renderer',
-						initializeRenderer: () => renderer,
-					}),
-				],
-			} as EcoPagesAppConfig,
-			assetProcessingService: AssetService,
-			runtimeOrigin: 'http://localhost:3000',
-		});
-		renderer.MockComponentRenderResult = {
-			html: '<aside>Local renderer</aside>',
-			canAttachAttributes: true,
-			rootTag: 'aside',
-			integrationName: 'test-renderer',
-		};
-
-		const ForeignComponent = (() => '<aside>Foreign</aside>') as EcoComponent<Record<string, unknown>>;
-		ForeignComponent.config = {
-			integration: 'foreign-renderer',
-			__eco: {
-				id: 'foreign-component',
-				file: '/app/components/foreign-component.tsx',
-				integration: 'foreign-renderer',
-			},
-		};
-
-		const rendererCache = new Map<string, IntegrationRenderer<any>>();
-		const result = await renderer.renderComponentWithForeignChildren({
-			component: ForeignComponent,
-			props: { label: 'foreign' },
-			integrationContext: { rendererCache },
-		});
-
-		expect(result.html).toBe('<aside>Local renderer</aside>');
-	});
-
 	describe('renderToResponse', () => {
 		it('should render a view with default status 200', async () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1050,8 +326,8 @@ describe('IntegrationRenderer', () => {
 
 		it('should render a partial view without layout', async () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1068,8 +344,8 @@ describe('IntegrationRenderer', () => {
 
 		it('should apply custom status code', async () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1082,8 +358,8 @@ describe('IntegrationRenderer', () => {
 
 		it('should apply custom headers', async () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1106,8 +382,8 @@ describe('IntegrationRenderer', () => {
 
 		it('should render with layout when not partial', async () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1130,8 +406,8 @@ describe('IntegrationRenderer', () => {
 	describe('renderComponent', () => {
 		it('should render a component with structured output', async () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1152,8 +428,8 @@ describe('IntegrationRenderer', () => {
 	describe('execute component-level artifacts', () => {
 		it('should handle stream-like render bodies without re-consuming disturbed responses', async () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1175,10 +451,6 @@ describe('IntegrationRenderer', () => {
 		});
 
 		it('should apply document attributes during route finalization', async () => {
-			const appConfig = {
-				...AppConfig,
-			} as EcoPagesAppConfig;
-
 			class DocumentAttributeRenderer extends TestIntegrationRenderer {
 				protected override getDocumentAttributes(): Record<string, string> | undefined {
 					return { 'data-eco-document-owner': 'react-router' };
@@ -1186,8 +458,8 @@ describe('IntegrationRenderer', () => {
 			}
 
 			const renderer = new DocumentAttributeRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1211,7 +483,7 @@ describe('IntegrationRenderer', () => {
 
 		it('should not force render nested dependency components without resolved props context', async () => {
 			const explicitRenderer = {
-				renderComponent: vi.fn(async () => ({
+				renderComponent: vi.fn(async (_input: ComponentRenderInput) => ({
 					html: '<aside>Nested</aside>',
 					canAttachAttributes: true,
 					rootTag: 'aside',
@@ -1220,10 +492,10 @@ describe('IntegrationRenderer', () => {
 				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) =>
 					explicitRenderer.renderComponent(input),
 				),
-			} as unknown as IntegrationRenderer;
+			};
 
 			const appConfig = {
-				...AppConfig,
+				...testAppConfig,
 				integrations: [
 					createMockIntegrationPlugin({
 						name: 'explicit-renderer',
@@ -1234,7 +506,7 @@ describe('IntegrationRenderer', () => {
 
 			const renderer = new TestIntegrationRenderer({
 				appConfig,
-				assetProcessingService: AssetService,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -1276,850 +548,21 @@ describe('IntegrationRenderer', () => {
 				query: {},
 			});
 
-			expect((explicitRenderer.renderComponent as any).mock.calls).toHaveLength(0);
+			expect(explicitRenderer.renderComponent).toHaveBeenCalledTimes(0);
 
 			const body = await new Response(result.body as BodyInit).text();
 			expect(body).toContain('<main>Test Page</main>');
 
-			const processedDeps = (renderer as any).htmlTransformer.getProcessedDependencies();
+			const processedDeps = renderer.getTestProcessedDependencies();
 			expect(processedDeps.some((dep: ProcessedAsset) => dep.srcUrl === '/assets/nested-explicit.js')).toBe(
 				false,
 			);
 		});
 
-		it('should include global integration dependencies for referenced component integrations once', async () => {
-			const explicitIntegrationDependency = {
-				kind: 'script',
-				srcUrl: '/assets/react-runtime.js',
-				position: 'head',
-			} as ProcessedAsset;
-
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'react',
-						initializeRenderer: () => renderer as unknown as IntegrationRenderer,
-						getResolvedIntegrationDependencies: () => [explicitIntegrationDependency],
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			renderer.RenderedBody = '<html><body><main>Test Page</main></body></html>';
-			renderer.MockComponentRenderResult = {
-				html: '<main>Test Page</main>',
-				canAttachAttributes: true,
-				rootTag: 'main',
-				integrationName: 'test-renderer',
-			};
-
-			const ReactNested = (() => '<div>React Nested</div>') as EcoComponent<Record<string, unknown>>;
-			ReactNested.config = {
-				integration: 'react',
-				__eco: {
-					id: 'react-nested',
-					file: '/app/components/react-nested.react.tsx',
-					integration: 'react',
-				},
-			};
-
-			const Page = (() => '<main>Test Page</main>') as unknown as EcoPageComponent<any>;
-			Page.config = {
-				dependencies: {
-					components: [ReactNested],
-				},
-			};
-
-			renderer.PageModule = {
-				default: Page,
-			};
-			renderer.HtmlTemplate = (() =>
-				'<html><body><main>Test Page</main></body></html>') as EcoComponent<HtmlTemplateProps>;
-
-			await renderer.execute({
-				file: '/app/pages/index.ts',
-				params: {},
-				query: {},
-			});
-
-			const processedDeps = (renderer as any).htmlTransformer.getProcessedDependencies();
-			expect(
-				processedDeps.filter((dep: ProcessedAsset) => dep.srcUrl === '/assets/react-runtime.js'),
-			).toHaveLength(1);
-		});
-
-		it('should fail route execution when unresolved eco-marker artifact HTML is returned', async () => {
-			const explicitRenderer = {
-				renderComponent: vi.fn(async () => ({
-					html: '<aside>Nested Render</aside>',
-					canAttachAttributes: true,
-					rootTag: 'aside',
-					integrationName: 'explicit-renderer',
-					rootAttributes: { 'data-eco-component-id': 'nested-1' },
-					assets: [
-						{
-							kind: 'script',
-							srcUrl: '/assets/nested-render.js',
-							position: 'head',
-						} as ProcessedAsset,
-					],
-				})),
-				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) =>
-					explicitRenderer.renderComponent(input),
-				),
-			} as unknown as IntegrationRenderer;
-
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'explicit-renderer',
-						initializeRenderer: () => explicitRenderer,
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			renderer.RenderedBody =
-				'<html><body><main>Before<eco-marker data-eco-node-id="n_1" data-eco-component-ref="nested-component" data-eco-props-ref="props-1"></eco-marker>After</main></body></html>';
-			renderer.MockComponentRenderResult = {
-				html: '<main>Test Page</main>',
-				canAttachAttributes: true,
-				rootTag: 'main',
-				integrationName: 'test-renderer',
-			};
-
-			const NestedComponent = (() => '<aside>Nested</aside>') as EcoComponent<Record<string, unknown>>;
-			NestedComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'nested-component',
-					file: '/app/components/nested-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const Page = (() => '<main>Test Page</main>') as unknown as EcoPageComponent<any>;
-			Page.config = {
-				dependencies: {
-					components: [NestedComponent],
-				},
-			};
-
-			renderer.PageModule = {
-				default: Page,
-			} as unknown as EcoPageFile;
-			renderer.RenderBodyFactory = () => renderer.RenderedBody;
-			renderer.HtmlTemplate = (() =>
-				'<html><body><main>Test Page</main></body></html>') as EcoComponent<HtmlTemplateProps>;
-
-			await expect(
-				renderer.execute({
-					file: '/app/pages/index.ts',
-					params: {},
-					query: {},
-				}),
-			).rejects.toThrow('Full-route unresolved-marker fallback has been removed');
-			expect((explicitRenderer.renderComponent as any).mock.calls).toHaveLength(0);
-		});
-
-		it('should fail route execution when unresolved eco-marker artifact HTML remains inside surrounding shell html', async () => {
-			const explicitRenderer = {
-				renderComponent: vi.fn(async () => ({
-					html: '<aside data-explicit-shell="nested"><span>Nested Render</span></aside>',
-					canAttachAttributes: true,
-					rootTag: 'aside',
-					integrationName: 'explicit-renderer',
-					rootAttributes: { 'data-eco-component-id': 'nested-contract-root' },
-					assets: [
-						{
-							kind: 'script',
-							srcUrl: '/assets/explicit-contract.js',
-							position: 'head',
-						} as ProcessedAsset,
-					],
-				})),
-				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) =>
-					explicitRenderer.renderComponent(input),
-				),
-			} as unknown as IntegrationRenderer;
-
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'explicit-renderer',
-						initializeRenderer: () => explicitRenderer,
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			renderer.RenderedBody =
-				'<html><body><main><section data-shell="outer">Before<eco-marker data-eco-node-id="n_1" data-eco-component-ref="nested-component" data-eco-props-ref="props-1"></eco-marker>After</section></main></body></html>';
-			renderer.MockComponentRenderResult = {
-				html: '<main>Test Page</main>',
-				canAttachAttributes: true,
-				rootTag: 'main',
-				integrationName: 'test-renderer',
-			};
-
-			const NestedComponent = (() => '<aside>Nested</aside>') as EcoComponent<Record<string, unknown>>;
-			NestedComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'nested-component',
-					file: '/app/components/nested-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const Page = (() => '<main>Test Page</main>') as unknown as EcoPageComponent<any>;
-			Page.config = {
-				dependencies: {
-					components: [NestedComponent],
-				},
-			};
-
-			renderer.PageModule = {
-				default: Page,
-			} as unknown as EcoPageFile;
-			renderer.RenderBodyFactory = () => renderer.RenderedBody;
-			renderer.HtmlTemplate = (() =>
-				'<html><body><main>Test Page</main></body></html>') as EcoComponent<HtmlTemplateProps>;
-
-			await expect(
-				renderer.execute({
-					file: '/app/pages/index.ts',
-					params: {},
-					query: {},
-				}),
-			).rejects.toThrow('Full-route unresolved-marker fallback has been removed');
-			expect((explicitRenderer.renderComponent as any).mock.calls).toHaveLength(0);
-		});
-
-		it('should fail route execution for deep multi-level unresolved eco-marker artifacts', async () => {
-			const renderOrder: string[] = [];
-			const explicitRenderer = {
-				renderComponent: vi.fn(async (input: ComponentRenderInput) => {
-					const componentId = input.component.config?.__eco?.id;
-					renderOrder.push(componentId as string);
-
-					if (componentId === 'leaf-component') {
-						return {
-							html: '<span>Leaf Render</span>',
-							canAttachAttributes: true,
-							rootTag: 'span',
-							integrationName: 'explicit-renderer',
-						};
-					}
-
-					if (componentId === 'parent-component') {
-						return {
-							html: `<section>${input.children ?? ''}</section>`,
-							canAttachAttributes: true,
-							rootTag: 'section',
-							integrationName: 'explicit-renderer',
-						};
-					}
-
-					return {
-						html: `<article>${input.children ?? ''}</article>`,
-						canAttachAttributes: true,
-						rootTag: 'article',
-						integrationName: 'explicit-renderer',
-						rootAttributes: { 'data-eco-component-id': 'root-node' },
-					};
-				}),
-				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) =>
-					explicitRenderer.renderComponent(input),
-				),
-			} as unknown as IntegrationRenderer;
-
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'explicit-renderer',
-						initializeRenderer: () => explicitRenderer,
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			const _parentMarker = createUnresolvedMarkerArtifact('n_2', 'parent-component', 'props-parent');
-			const _leafMarker = createUnresolvedMarkerArtifact('n_3', 'leaf-component', 'props-leaf');
-
-			renderer.RenderedBody = `<html><body><main>${createUnresolvedMarkerArtifact('n_1', 'root-component', 'props-root')}</main></body></html>`;
-			renderer.MockComponentRenderResult = {
-				html: '<main>Test Page</main>',
-				canAttachAttributes: true,
-				rootTag: 'main',
-				integrationName: 'test-renderer',
-			};
-
-			const RootComponent = (() => '<article>Root</article>') as EcoComponent<Record<string, unknown>>;
-			RootComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'root-component',
-					file: '/app/components/root-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const ParentComponent = (() => '<section>Parent</section>') as EcoComponent<Record<string, unknown>>;
-			ParentComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'parent-component',
-					file: '/app/components/parent-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const LeafComponent = (() => '<span>Leaf</span>') as EcoComponent<Record<string, unknown>>;
-			LeafComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'leaf-component',
-					file: '/app/components/leaf-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const Page = (() => '<main>Test Page</main>') as unknown as EcoPageComponent<any>;
-			Page.config = {
-				dependencies: {
-					components: [RootComponent, ParentComponent, LeafComponent],
-				},
-			};
-
-			renderer.PageModule = {
-				default: Page,
-			} as unknown as EcoPageFile;
-			renderer.RenderBodyFactory = () => renderer.RenderedBody;
-			renderer.HtmlTemplate = (() =>
-				'<html><body><main>Test Page</main></body></html>') as EcoComponent<HtmlTemplateProps>;
-
-			await expect(
-				renderer.execute({
-					file: '/app/pages/index.ts',
-					params: {},
-					query: {},
-				}),
-			).rejects.toThrow('Full-route unresolved-marker fallback has been removed');
-			expect(renderOrder).toEqual([]);
-		});
-
-		it('should fail route execution when props reference is missing', async () => {
-			const explicitRenderer = {
-				renderComponent: vi.fn(async () => ({
-					html: '<aside>Nested Render</aside>',
-					canAttachAttributes: true,
-					rootTag: 'aside',
-					integrationName: 'explicit-renderer',
-				})),
-				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) =>
-					explicitRenderer.renderComponent(input),
-				),
-			} as unknown as IntegrationRenderer;
-
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'explicit-renderer',
-						initializeRenderer: () => explicitRenderer,
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			renderer.RenderedBody =
-				'<html><body><main><eco-marker data-eco-node-id="n_1" data-eco-component-ref="nested-component" data-eco-props-ref="props-missing"></eco-marker></main></body></html>';
-			renderer.MockComponentRenderResult = {
-				html: '<main>Test Page</main>',
-				canAttachAttributes: true,
-				rootTag: 'main',
-				integrationName: 'test-renderer',
-			};
-
-			const NestedComponent = (() => '<aside>Nested</aside>') as EcoComponent<Record<string, unknown>>;
-			NestedComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'nested-component',
-					file: '/app/components/nested-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const Page = (() => '<main>Test Page</main>') as unknown as EcoPageComponent<any>;
-			Page.config = {
-				dependencies: {
-					components: [NestedComponent],
-				},
-			};
-
-			renderer.PageModule = {
-				default: Page,
-			} as unknown as EcoPageFile;
-			renderer.HtmlTemplate = (() =>
-				'<html><body><main>Test Page</main></body></html>') as EcoComponent<HtmlTemplateProps>;
-
-			await expect(
-				renderer.execute({
-					file: '/app/pages/index.ts',
-					params: {},
-					query: {},
-				}),
-			).rejects.toThrow('Full-route unresolved-marker fallback has been removed');
-		});
-
-		it('should not recursively resolve unresolved eco-marker artifacts that were only passed through resolved child html', async () => {
-			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			renderer.MockComponentRenderResult = {
-				html: '<section><eco-marker data-eco-node-id="n_2" data-eco-component-ref="nested-component" data-eco-props-ref="p_2"></eco-marker></section>',
-				canAttachAttributes: true,
-				rootTag: 'section',
-				integrationName: 'test-renderer',
-			};
-
-			const Component = (() => '<section />') as EcoComponent<Record<string, unknown>>;
-			Component.config = {
-				integration: 'test-renderer',
-				__eco: {
-					id: 'component',
-					file: '/app/components/component.ts',
-					integration: 'test-renderer',
-				},
-			};
-
-			await expect(
-				renderer.renderComponentWithForeignChildren({
-					component: Component,
-					props: {},
-					children: '<aside>already resolved child html</aside>',
-				}),
-			).resolves.toEqual(
-				expect.objectContaining({
-					html: '<section><eco-marker data-eco-node-id="n_2" data-eco-component-ref="nested-component" data-eco-props-ref="p_2"></eco-marker></section>',
-				}),
-			);
-		});
-
-		it('fails fast when a renderer without a foreign-child runtime crosses into a foreign owner', async () => {
-			const foreignRenderer = {
-				renderComponent: vi.fn(async () => ({
-					html: '<span>resolved nested marker</span>',
-					canAttachAttributes: true,
-					rootTag: 'span',
-					integrationName: 'foreign-renderer',
-				})),
-				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) =>
-					foreignRenderer.renderComponent(input),
-				),
-			} as unknown as IntegrationRenderer;
-
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'foreign-renderer',
-						initializeRenderer: () => foreignRenderer,
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-			renderer.UseFailFastForeignChildRuntime = true;
-
-			const ForeignComponent = eco.component<{}, string>({
-				integration: 'foreign-renderer',
-				render: () => '<span>foreign</span>',
-			});
-
-			const ShellComponent = eco.component<{ children?: string }, string>({
-				integration: 'test-renderer',
-				dependencies: {
-					components: [ForeignComponent],
-				},
-				render: ({ children }) => `<section>${children ?? ''}${ForeignComponent({})}</section>`,
-			});
-
-			const passedThroughMarker = createUnresolvedMarkerArtifact(
-				'n_passed',
-				'passed-through-component',
-				'p_passed',
-			);
-
-			await expect(
-				renderer.renderComponentWithForeignChildren({
-					component: ShellComponent,
-					props: { children: passedThroughMarker },
-					children: passedThroughMarker,
-				}),
-			).rejects.toThrow('without a renderer-owned foreign-child runtime');
-			expect(foreignRenderer.renderComponent).toHaveBeenCalledTimes(0);
-		});
-
-		it('fails route execution when deep mixed-integration eco-marker artifacts are returned at the route level', async () => {
-			const renderOrder: string[] = [];
-			const explicitRenderer = {
-				renderComponent: vi.fn(async (input: ComponentRenderInput) => {
-					const componentId = input.component.config?.__eco?.id as string | undefined;
-					if (componentId) {
-						renderOrder.push(componentId);
-					}
-
-					if (componentId === 'leaf-component') {
-						return {
-							html: '<span data-leaf="true">Leaf Render</span>',
-							canAttachAttributes: true,
-							rootTag: 'span',
-							integrationName: 'explicit-renderer',
-						};
-					}
-
-					if (componentId === 'parent-component') {
-						return {
-							html: `<section data-parent="true">${input.children ?? ''}</section>`,
-							canAttachAttributes: true,
-							rootTag: 'section',
-							integrationName: 'explicit-renderer',
-						};
-					}
-
-					return {
-						html: `<article data-root="true">${input.children ?? ''}</article>`,
-						canAttachAttributes: true,
-						rootTag: 'article',
-						integrationName: 'explicit-renderer',
-					};
-				}),
-				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) =>
-					explicitRenderer.renderComponent(input),
-				),
-			} as unknown as IntegrationRenderer;
-
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'explicit-renderer',
-						initializeRenderer: () => explicitRenderer,
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			const _parentMarker = createUnresolvedMarkerArtifact('n_2', 'parent-component', 'props-parent');
-			const _leafMarker = createUnresolvedMarkerArtifact('n_3', 'leaf-component', 'props-leaf');
-
-			renderer.RenderedBody = `<html><body><main><div data-shell="deep">${createUnresolvedMarkerArtifact('n_1', 'root-component', 'props-root')}</div></main></body></html>`;
-			renderer.MockComponentRenderResult = {
-				html: '<main>Test Page</main>',
-				canAttachAttributes: true,
-				rootTag: 'main',
-				integrationName: 'test-renderer',
-			};
-
-			const RootComponent = (() => '<article>Root</article>') as EcoComponent<Record<string, unknown>>;
-			RootComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'root-component',
-					file: '/app/components/root-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const ParentComponent = (() => '<section>Parent</section>') as EcoComponent<Record<string, unknown>>;
-			ParentComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'parent-component',
-					file: '/app/components/parent-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const LeafComponent = (() => '<span>Leaf</span>') as EcoComponent<Record<string, unknown>>;
-			LeafComponent.config = {
-				integration: 'explicit-renderer',
-				__eco: {
-					id: 'leaf-component',
-					file: '/app/components/leaf-component.ts',
-					integration: 'explicit-renderer',
-				},
-			};
-
-			const Page = (() => '<main>Test Page</main>') as unknown as EcoPageComponent<any>;
-			Page.config = {
-				dependencies: {
-					components: [RootComponent, ParentComponent, LeafComponent],
-				},
-			};
-
-			renderer.PageModule = {
-				default: Page,
-			} as unknown as EcoPageFile;
-			renderer.RenderBodyFactory = () => renderer.RenderedBody;
-			renderer.HtmlTemplate = (() =>
-				'<html><body><main>Test Page</main></body></html>') as EcoComponent<HtmlTemplateProps>;
-
-			await expect(
-				renderer.execute({
-					file: '/app/pages/index.ts',
-					params: {},
-					query: {},
-				}),
-			).rejects.toThrow('Full-route unresolved-marker fallback has been removed');
-			expect(renderOrder).toEqual([]);
-		});
-
-		it('renders same-integration leaf components under their own integration context', async () => {
-			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			const LeafComponent = eco.component<{}, string>({
-				integration: 'test-renderer',
-				render: () => '<section>Leaf Render</section>',
-			});
-
-			const result = await runWithComponentRenderContext(
-				{
-					currentIntegration: 'foreign-renderer',
-				},
-				async () =>
-					renderer.renderComponentWithForeignChildren({
-						component: LeafComponent,
-						props: {},
-					}),
-			);
-
-			expect(result.value.html).toBe('<section>Leaf Render</section>');
-			expect(result.value.html).not.toContain('<eco-marker');
-			expect(renderer.ForeignChildRuntimeCreationCount).toBe(0);
-		});
-
-		it('uses inline partial rendering when no foreign children are present', async () => {
-			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-			const View = (() => '<section>Inline</section>') as EcoComponent<Record<string, unknown>>;
-			View.config = {
-				integration: 'test-renderer',
-				__eco: {
-					id: 'inline-view',
-					file: '/app/components/inline-view.ts',
-					integration: 'test-renderer',
-				},
-			};
-
-			let inlineRenderCount = 0;
-			const response = await renderer.testRenderPartialViewResponse({
-				view: View,
-				props: {},
-				renderInline: async () => {
-					inlineRenderCount += 1;
-					return '<section>Inline</section>';
-				},
-			});
-
-			expect(await response.text()).toBe('<section>Inline</section>');
-			expect(inlineRenderCount).toBe(1);
-			expect(renderer.ForeignChildRenderCount).toBe(0);
-		});
-
-		it('falls back to foreign-subtree partial rendering when compatibility is needed', async () => {
-			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-			const ForeignChild = (() => '<span>Foreign</span>') as EcoComponent<Record<string, unknown>>;
-			ForeignChild.config = {
-				integration: 'react',
-				__eco: {
-					id: 'foreign-child',
-					file: '/app/components/foreign-child.tsx',
-					integration: 'react',
-				},
-			};
-
-			const View = (() => '<section>Foreign Subtree</section>') as EcoComponent<Record<string, unknown>>;
-			View.config = {
-				integration: 'test-renderer',
-				__eco: {
-					id: 'foreign-subtree-view',
-					file: '/app/components/foreign-subtree-view.ts',
-					integration: 'test-renderer',
-				},
-				dependencies: { components: [ForeignChild] },
-			};
-			renderer.MockComponentRenderResult = {
-				html: '<section>Foreign Subtree</section>',
-				canAttachAttributes: true,
-				rootTag: 'section',
-				integrationName: 'test-renderer',
-			};
-
-			let inlineRenderCount = 0;
-			const response = await renderer.testRenderPartialViewResponse({
-				view: View,
-				props: {},
-				renderInline: async () => {
-					inlineRenderCount += 1;
-					return '<section>Inline</section>';
-				},
-			});
-
-			expect(await response.text()).toBe('<section>Foreign Subtree</section>');
-			expect(inlineRenderCount).toBe(0);
-			expect(renderer.ForeignChildRenderCount).toBe(1);
-		});
-
-		it('reuses one foreign renderer instance across shared view shell composition', async () => {
-			const foreignRenderer = {
-				renderComponentWithForeignChildren: vi.fn(async (input: ComponentRenderInput) => {
-					const componentId = input.component.config?.__eco?.id;
-
-					if (componentId === 'foreign-html-template') {
-						return {
-							html: `<html><body>${String(input.children ?? '')}</body></html>`,
-							canAttachAttributes: true,
-							rootTag: 'html',
-							integrationName: 'foreign-renderer',
-						};
-					}
-
-					if (componentId === 'foreign-layout') {
-						return {
-							html: `<main>${String(input.children ?? '')}</main>`,
-							canAttachAttributes: true,
-							rootTag: 'main',
-							integrationName: 'foreign-renderer',
-						};
-					}
-
-					return {
-						html: '<section>Foreign View</section>',
-						canAttachAttributes: true,
-						rootTag: 'section',
-						integrationName: 'foreign-renderer',
-					};
-				}),
-			} as unknown as IntegrationRenderer;
-
-			const initializeRenderer = vi.fn(() => foreignRenderer);
-			const appConfig = {
-				...AppConfig,
-				integrations: [
-					createMockIntegrationPlugin({
-						name: 'foreign-renderer',
-						initializeRenderer,
-					}),
-				],
-			} as unknown as EcoPagesAppConfig;
-
-			const renderer = new TestIntegrationRenderer({
-				appConfig,
-				assetProcessingService: AssetService,
-				runtimeOrigin: 'http://localhost:3000',
-			});
-
-			const View = (() => '<section>Foreign View</section>') as EcoComponent<Record<string, unknown>>;
-			View.config = {
-				integration: 'foreign-renderer',
-				__eco: {
-					id: 'foreign-view',
-					file: '/app/components/foreign-view.ts',
-					integration: 'foreign-renderer',
-				},
-			};
-
-			const Layout = (() => '<main />') as EcoComponent<Record<string, unknown>>;
-			Layout.config = {
-				integration: 'foreign-renderer',
-				__eco: {
-					id: 'foreign-layout',
-					file: '/app/components/foreign-layout.ts',
-					integration: 'foreign-renderer',
-				},
-			};
-
-			renderer.HtmlTemplate = (() => '<html><body></body></html>') as EcoComponent<HtmlTemplateProps>;
-			renderer.HtmlTemplate.config = {
-				integration: 'foreign-renderer',
-				__eco: {
-					id: 'foreign-html-template',
-					file: '/app/components/foreign-html-template.ts',
-					integration: 'foreign-renderer',
-				},
-			};
-
-			const response = await renderer.testRenderViewWithDocumentShell({
-				view: View,
-				props: {},
-				layout: Layout,
-			});
-
-			expect(await response.text()).toContain('<main><section>Foreign View</section></main>');
-			expect(initializeRenderer).toHaveBeenCalledTimes(1);
-			expect(foreignRenderer.renderComponentWithForeignChildren).toHaveBeenCalledTimes(3);
-		});
-
 		it('should skip foreign-child wrapping for pure same-integration component trees', () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
 
@@ -2149,8 +592,8 @@ describe('IntegrationRenderer', () => {
 
 		it('should detect nested cross-integration component trees', () => {
 			const renderer = new TestIntegrationRenderer({
-				appConfig: AppConfig,
-				assetProcessingService: AssetService,
+				appConfig: testAppConfig,
+				assetProcessingService: testAssetService,
 				runtimeOrigin: 'http://localhost:3000',
 			});
