+++ packages/core/src/route-renderer/orchestration/page-browser-graph/page-browser-graph-merge.utils.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { createPagePackage, type ProcessedAsset } from '../../../services/assets/asset-processing-service/index.ts';
-import { mergePageBrowserGraphIntoPagePackage } from './page-browser-graph-merge.utils.ts';
+import { mergePageBrowserGraph } from './page-browser-graph-merge.utils.ts';
 
 function scriptAsset(filepath: string): ProcessedAsset {
 	return {
@@ -11,60 +11,68 @@ function scriptAsset(filepath: string): ProcessedAsset {
 	} as ProcessedAsset;
 }
 
-describe('mergePageBrowserGraphIntoPagePackage', () => {
-	it('returns undefined when no page browser graph is provided', () => {
-		const host = {
-			getPagePackage: () => undefined,
-			getProcessedDependencies: () => [],
-			dedupeProcessedAssets: (assets: readonly ProcessedAsset[]) => [...assets],
-			setPagePackage: () => {
-				throw new Error('should not set package');
-			},
-		};
-
-		expect(mergePageBrowserGraphIntoPagePackage(host, undefined)).toBeUndefined();
-	});
-
+describe('mergePageBrowserGraph', () => {
 	it('merges graph assets into an existing page package and rebuilds the package', () => {
 		const existing = scriptAsset('/existing.js');
 		const entry = scriptAsset('/entry.js');
 		const chunk = scriptAsset('/chunk.js');
-		let stored = createPagePackage([existing], {
+		const current = createPagePackage([existing], {
 			pageBrowserGraph: {
 				entryAssets: [existing],
 				chunkAssets: [],
 			},
 		});
 
-		const host = {
-			getPagePackage: () => stored,
-			getProcessedDependencies: () => [existing],
-			dedupeProcessedAssets: (assets: readonly ProcessedAsset[]) => {
-				const seen = new Set<string>();
-				return assets.filter((asset) => {
-					const key = String((asset as { filepath?: string }).filepath ?? '');
-					if (seen.has(key)) {
-						return false;
-					}
-					seen.add(key);
-					return true;
-				});
-			},
-			setPagePackage: (pagePackage: typeof stored) => {
-				stored = pagePackage;
-			},
-		};
-
-		const merged = mergePageBrowserGraphIntoPagePackage(host, {
+		const { pagePackage, mergedGraph } = mergePageBrowserGraph(current, [existing], {
 			entryAssets: [entry],
 			chunkAssets: [chunk],
 		});
 
-		expect(merged?.entryAssets.map((asset) => (asset as { filepath?: string }).filepath)).toEqual([
+		expect(mergedGraph.entryAssets.map((asset) => (asset as { filepath?: string }).filepath)).toEqual([
 			'/existing.js',
 			'/entry.js',
 		]);
-		expect(merged?.chunkAssets.map((asset) => (asset as { filepath?: string }).filepath)).toEqual(['/chunk.js']);
-		expect(stored.pageBrowserGraph).toEqual(merged);
+		expect(mergedGraph.chunkAssets.map((asset) => (asset as { filepath?: string }).filepath)).toEqual([
+			'/chunk.js',
+		]);
+		expect(pagePackage.pageBrowserGraph).toEqual(mergedGraph);
+	});
+
+	it('uses processed dependencies when no current page package exists', () => {
+		const base = scriptAsset('/base.js');
+		const entry = scriptAsset('/entry.js');
+
+		const { pagePackage, mergedGraph } = mergePageBrowserGraph(undefined, [base], {
+			entryAssets: [entry],
+			chunkAssets: [],
+		});
+
+		expect(mergedGraph.entryAssets).toEqual([entry]);
+		expect(pagePackage.assets.map((asset) => (asset as { filepath?: string }).filepath)).toEqual([
+			'/base.js',
+			'/entry.js',
+		]);
+		expect(pagePackage.pageBrowserGraph).toEqual(mergedGraph);
+	});
+
+	it('preserves the page browser graph when package assets are rebuilt', () => {
+		const entry = scriptAsset('/entry.js');
+		const chunk = scriptAsset('/chunk.js');
+		const extra = scriptAsset('/extra.js');
+		const current = createPagePackage([extra], {
+			pageBrowserGraph: {
+				entryAssets: [entry],
+				chunkAssets: [chunk],
+			},
+		});
+
+		const { pagePackage, mergedGraph } = mergePageBrowserGraph(current, [extra], {
+			entryAssets: [entry],
+			chunkAssets: [chunk],
+		});
+
+		expect(pagePackage.pageBrowserGraph).toEqual(mergedGraph);
+		expect(mergedGraph.entryAssets).toEqual([entry]);
+		expect(mergedGraph.chunkAssets).toEqual([chunk]);
 	});
 });
