+++ packages/integrations/lit/src/lit-renderer.ts
@@ -21,7 +21,14 @@ import {
 	IntegrationRenderer,
 	type RenderToResponseContext,
 } from '@ecopages/core/route-renderer/orchestration/integration-renderer';
-import type { QueuedForeignSubtreeResolutionContext } from '@ecopages/core/route-renderer/orchestration/foreign-child/foreign-subtree-execution.service';
+import type {
+	ForeignSubtreeExecutionOwningRenderer,
+	QueuedForeignSubtreeResolutionContext,
+} from '@ecopages/core/route-renderer/orchestration/foreign-child/foreign-subtree-execution.service';
+import {
+	getForeignSubtreeResolutionContextKey,
+	resolveOwningIntegrationRenderer,
+} from '@ecopages/core/route-renderer/orchestration/foreign-child/owning-renderer-resolution';
 import { LitSsrLazyPreloader } from './lit-ssr-lazy-preloader.ts';
 import type { LitStaticRenderSession } from './lit-static-render-session.ts';
 import { LIT_PLUGIN_NAME } from './lit.constants.ts';
@@ -174,13 +181,30 @@ export class LitRenderer extends IntegrationRenderer<EcoPagesElement> {
 		const renderedHtml = await renderLitValueToString(content);
 		const html =
 			renderedChildren === undefined ? renderedHtml : injectLitRenderedChildren(renderedHtml, renderedChildren);
-		const queuedForeignSubtreeResolution = await this.resolveQueuedForeignSubtreeHtml(
+		const queuedForeignSubtreeResolution = await this.foreignSubtreeExecutionService.resolveQueuedHtml({
+			currentIntegrationName: this.name,
 			html,
-			this.getQueuedForeignSubtreeResolutionContext<QueuedForeignSubtreeResolutionContext>(input),
-			(children, _runtimeContext, queuedResolutionsByToken, resolveToken) =>
+			runtimeContext:
+				this.foreignSubtreeExecutionService.getQueuedRuntimeContext<QueuedForeignSubtreeResolutionContext>(
+					input,
+					getForeignSubtreeResolutionContextKey(this.name),
+				),
+			queueLabel: 'Lit',
+			getOwningRenderer: (integrationName, rendererCache) =>
+				resolveOwningIntegrationRenderer({
+					appConfig: this.appConfig,
+					runtimeOrigin: this.runtimeOrigin,
+					currentIntegrationName: this.name,
+					currentRenderer: this,
+					integrationName,
+					cache: rendererCache as Map<string, ForeignSubtreeExecutionOwningRenderer>,
+				}),
+			applyAttributesToFirstElement: (resolvedHtml, attributes) =>
+				this.htmlTransformer.applyAttributesToFirstElement(resolvedHtml, attributes),
+			dedupeProcessedAssets: (assets) => this.htmlTransformer.dedupeProcessedAssets(assets),
+			renderQueuedChildren: (children, _runtimeContext, queuedResolutionsByToken, resolveToken) =>
 				this.renderLitQueuedForeignSubtreeChildren(children, queuedResolutionsByToken, resolveToken),
-			'Lit',
-		);
+		});
 		const hasDependencies = Boolean(input.component.config?.dependencies);
 		const canResolveAssets = typeof this.assetProcessingService?.processDependencies === 'function';
 		const assets =
